requirements.txt
discord.py
requests
praw
COVID19Py
psycopg2
wikipedia
pyowm